+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Simulation
 
   * The LCP solver will be less aggressive about printing out unnecessary warnings: [#1238](https://github.com/dartsim/dart/pull/1238)
+  * Fixed not copying constraints in World::setConstraintSolver(): [#1260](https://github.com/dartsim/dart/pull/1260)
 
 * Collision Detection